+++ packages/radix-ui-themes/src/components/checkbox.props.ts
@@ -2,7 +2,7 @@ import { colorProp, highContrastProp } from '../helpers';
 import type { PropDef } from '../helpers';
 
 const sizes = ['1', '2'] as const;
-const variants = ['solid', 'surface', 'soft'] as const;
+const variants = ['solid', 'classic', 'soft'] as const;
 
 const checkboxPropDefs = {
   size: { type: 'enum', values: sizes, default: '1', responsive: true },