+++ packages/radix-ui-themes/src/components/text-area.props.ts
@@ -2,7 +2,7 @@ import { colorProp } from '../helpers';
 import { PropDef } from '../helpers';
 
 const sizes = ['1', '2', '3'] as const;
-const variants = ['solid', 'surface', 'soft'] as const;
+const variants = ['solid', 'classic', 'soft'] as const;
 
 const textAreaPropDefs = {
   size: { type: 'enum', values: sizes, default: '2', responsive: true },