+++ packages/radix-ui-themes/src/components/text-field.props.ts
@@ -3,7 +3,7 @@ import { flexPropDefs } from './flex.props';
 import { PropDef } from '../helpers';
 
 const sizes = ['1', '2', '3'] as const;
-const variants = ['solid', 'surface', 'soft'] as const;
+const variants = ['solid', 'classic', 'soft'] as const;
 
 const textFieldPropDefs = {
   size: { type: 'enum', values: sizes, default: '2', responsive: true },